+Ajout service de deletion des Releves en supprimant les Lignes de relevé associées. / +Ajout dial box de confirmation de suppression

+++ MainWindow.xaml.cs
@@ -23,14 +23,14 @@ namespace DataReporting
 		{
 			InitializeComponent();
 
+			ComboBoxSelectCapteur.ItemsSource = ServiceCapteur.GetCapteur().ToString();
 
 			ObservableCollection<BusinessCapteur> capteurs =  new ObservableCollection<BusinessCapteur>(ServiceCapteur.GetCapteur() as List<BusinessCapteur>);
-			gridTest.ItemsSource = capteurs;
-
-			//gridFromBdd.ItemsSource = ServiceLigneReleve.GetLignesReleve();
-			gridFromBdd.ItemsSource = ServiceLigneReleve.GetLignesReleve();
-			gridReleve.ItemsSource = ServiceReleve.GetReleve();
-			ComboBoxSelectCapteur.ItemsSource = ServiceCapteur.GetCapteur().ToString();
+			gridCapteur.ItemsSource = capteurs;
+			//gridReleve.ItemsSource = ServiceReleve.GetReleve();
+			//gridLigneReleve.ItemsSource = ServiceLigneReleve.GetLignesReleve();
+			
+			
 
 		}
 		
@@ -135,11 +135,13 @@ namespace DataReporting
 
 		private void SaveReleveToBDD(object sender, RoutedEventArgs e)
 		{
-			var cellInfo = gridTest.SelectedCells[0];
+			var cellInfo = gridCapteur.SelectedCells[0];
 			BusinessCapteur businessCapteur = (BusinessCapteur)cellInfo.Item;
 			int capteurId = businessCapteur.IdCapteur;
-			BusinessReleve businessReleve = new BusinessReleve();
-			businessReleve.CapteurID = capteurId;
+			BusinessReleve businessReleve = new BusinessReleve
+			{
+				CapteurID = capteurId
+			};
 
 			int idReleve = ServiceReleve.AddReleve(businessReleve);
 
@@ -148,27 +150,41 @@ namespace DataReporting
 			{
 				string[] list = Regex.Split(item.ToString(), @"\s+");
 
-				BusinessLigneReleve businessLigneReleve = new BusinessLigneReleve();
-				businessLigneReleve.DateLigneReleve = DateTime.Parse(list[1]);
-				businessLigneReleve.HeureLigneReleve = TimeSpan.Parse(list[2]);
-				businessLigneReleve.Temperature = double.Parse(list[3].Replace(".", ","));
-				businessLigneReleve.Hygrometrie = double.Parse(list[4].Remove(list[4].Length - 1).Replace(".", ","));
-				businessLigneReleve.ReleveID = idReleve;
+				BusinessLigneReleve businessLigneReleve = new BusinessLigneReleve
+				{
+					DateLigneReleve = DateTime.Parse(list[1]),
+					HeureLigneReleve = TimeSpan.Parse(list[2]),
+					Temperature = double.Parse(list[3].Replace(".", ",")),
+					Hygrometrie = double.Parse(list[4].Remove(list[4].Length - 1).Replace(".", ",")),
+					ReleveID = idReleve
+				};
 				lignesReleve.Add(businessLigneReleve);
 			}
 			ServiceLigneReleve.AddLignesReleve(lignesReleve);
 			MessageBox.Show("yeah");
+			listBoxReleve.Items.Clear();
+
 		}
 
 		// Capteur
-		private void DeleteTest_Click(object sender, RoutedEventArgs e)
+		private void BtnDeleteCapteur(object sender, RoutedEventArgs e)
 		{
-			var capteurARetirer = gridTest.SelectedItem as BusinessCapteur;
-			ServiceCapteur.DeleteCapteur(capteurARetirer);
-			gridTest.ItemsSource = ServiceCapteur.GetCapteur();
+			var capteurARetirer = gridCapteur.SelectedItem as BusinessCapteur;
+			string messageBoxText = "Voulez-vous supprimer le capteur selectionné?";
+			string caption = "Supprimer ?";
+			MessageBoxButton button = MessageBoxButton.YesNoCancel;
+			MessageBoxImage icon = MessageBoxImage.Question;
+			MessageBoxResult result = MessageBox.Show(messageBoxText, caption, button, icon);
+
+			if (result == MessageBoxResult.Yes)
+			{
+				ServiceCapteur.DeleteCapteur(capteurARetirer);
+				gridCapteur.ItemsSource = ServiceCapteur.GetCapteur();
+			}
+			
 		}
 
-		private void AddTest_click(object sender, RoutedEventArgs e)
+		private void BtnAjoutCapteur(object sender, RoutedEventArgs e)
 		{
 			BusinessCapteur nouveauCapteur = new BusinessCapteur
 			{
@@ -177,29 +193,65 @@ namespace DataReporting
 			};
 			ServiceCapteur.AddCapteur(nouveauCapteur);
 			MessageBox.Show("Capteur ajouté");
-			gridTest.ItemsSource = ServiceCapteur.GetCapteur();
+			gridCapteur.ItemsSource = ServiceCapteur.GetCapteur();
 		}
 
-		private void BtnSupprLigneCapteur(object sender, RoutedEventArgs e)
+		private void BtnSelectCapteur(object sender, RoutedEventArgs e)
 		{
+			BusinessCapteur businessCapteur = gridCapteur.SelectedItem as BusinessCapteur;
+			int capteurId = businessCapteur.IdCapteur;
+			MessageBox.Show("Capteur selectionné : Affichage liste des relevés du capteur.");
+			gridReleve.ItemsSource = ServiceReleve.GetReleveByCapteurId(capteurId);
 
-			var cellCapteur = gridTest.SelectedItem;
-			gridTest.SelectedCells.Clear();
-			//ItemsControl.ItemsSource
-			gridTest.Items.Remove(cellCapteur);
-			/*BusinessCapteur businessCapteur = (BusinessCapteur)cellCapteur.Item;
-			businessCapteur*/
+		}
+		//private void BtnSupprLigneCapteur(object sender, RoutedEventArgs e)
+		//{
+		//
+		//	//var cellCapteur = gridCapteur.SelectedItems;
+		//	
+		//	gridCapteur.SelectedItems.Clear();
+		//	/*BusinessCapteur businessCapteur = (BusinessCapteur)cellCapteur.Item;
+		//	businessCapteur*/
+		//}
+
+		// Releve
+		private void BtnSelectReleves(object sender, RoutedEventArgs e)
+		{
+			List<BusinessLigneReleve> releves = new List<BusinessLigneReleve>();
+			foreach (BusinessReleve businessReleve in gridReleve.SelectedItems) 
+			{
+				releves.AddRange(ServiceLigneReleve.GetLignesReleveById(businessReleve.IdReleve));
+			}
+			MessageBox.Show("Relevé(s) selectionné(s) : Affichage liste des Lignes.");
+			gridLigneReleve.ItemsSource = releves;
 		}
 
+
+		private void BtnSupprimerReleve(object sender, RoutedEventArgs e)
+		{
+			var refresh = gridReleve.SelectedItem as BusinessReleve;
+			var releveARetirer = gridReleve.SelectedItem as BusinessReleve;
+			string messageBoxText = "Voulez-vous supprimer le revelé selectionné?";
+			string caption = "Supprimer ?";
+			MessageBoxButton button = MessageBoxButton.YesNoCancel;
+			MessageBoxImage icon = MessageBoxImage.Question;
+			MessageBoxResult result = MessageBox.Show(messageBoxText, caption, button, icon);
+
+			if (result == MessageBoxResult.Yes)
+			{
+				ServiceReleve.DeleteReleve(releveARetirer);
+			}
+			gridLigneReleve.ItemsSource = null;
+			gridReleve.ItemsSource = ServiceReleve.GetReleveByCapteurId(releveARetirer.CapteurID);
+		}
 
 		// Test
-		private void BtnEnleverLigne_Click(object sender, RoutedEventArgs e)
+		private void BtnSupprimerLigneDeReleve(object sender, RoutedEventArgs e)
 		{
-			var releveARetirer = gridFromBdd.SelectedItem as BusinessReleve;
+			var releveARetirer = gridLigneReleve.SelectedItem as BusinessReleve;
 			ServiceLigneReleve.DeleteLigneReleve(releveARetirer);
-			gridFromBdd.ItemsSource = ServiceReleve.GetReleve();
+			gridLigneReleve.ItemsSource = ServiceReleve.GetReleve();
 		}
-
 
 	}
 }
